+++ packages/reactive_forms_annotations/lib/reactive_forms_annotations.dart
@@ -5,3 +5,4 @@ export 'src/form_group_annotation.dart';
 export 'src/form_array_annotation.dart';
 export 'src/form_control_annotation.dart';
 export 'src/form_model.dart';
+export 'src/extended_control.dart';