+++ packages/reactive_forms_generator/example/lib/docs/basic/basic_form.dart
@@ -29,6 +29,12 @@ class BasicFormWidget extends StatelessWidget {
                 ),
               ),
               const SizedBox(height: 8.0),
+              ReactiveValueListenableBuilder<double>(
+                formControlName: 'lightIntensity',
+                builder: (context, value, child) {
+                  return Container();
+                },
+              ),
               ReactiveTextField<String>(
                 formControl: formModel.passwordControl,
                 obscureText: true,