+++ packages/reactive_forms_generator/example/lib/docs/renamed_basic/renamed_basic.gform.dart
@@ -312,3 +312,65 @@ class ReactiveSomeWiredNameFormArrayBuilder<T> extends StatelessWidget {
     );
   }
 }
+
+class ReactiveSomeWiredNameFormFormGroupArrayBuilder<V>
+    extends StatelessWidget {
+  const ReactiveSomeWiredNameFormFormGroupArrayBuilder(
+      {Key? key,
+      this.extended,
+      this.getExtended,
+      this.builder,
+      required this.itemBuilder})
+      : assert(extended != null || getExtended != null,
+            "You have to specify `control` or `formControl`!"),
+        super(key: key);
+
+  final ExtendedControl<List<Map<String, Object?>?>, List<V>>? extended;
+
+  final ExtendedControl<List<Map<String, Object?>?>, List<V>> Function(
+      SomeWiredNameForm formModel)? getExtended;
+
+  final Widget Function(BuildContext context, List<Widget> itemList,
+      SomeWiredNameForm formModel)? builder;
+
+  final Widget Function(
+          BuildContext context, int i, V? item, SomeWiredNameForm formModel)
+      itemBuilder;
+
+  @override
+  Widget build(BuildContext context) {
+    final formModel = ReactiveSomeWiredNameForm.of(context);
+
+    if (formModel == null) {
+      throw FormControlParentNotFoundException(this);
+    }
+
+    final value = (extended ?? getExtended?.call(formModel))!;
+
+    return StreamBuilder<List<Map<String, Object?>?>?>(
+      stream: value.control.valueChanges,
+      builder: (context, snapshot) {
+        final itemList = (value.value() ?? <V>[])
+            .asMap()
+            .map((i, item) => MapEntry(
+                  i,
+                  itemBuilder(
+                    context,
+                    i,
+                    item,
+                    formModel,
+                  ),
+                ))
+            .values
+            .toList();
+
+        return builder?.call(
+              context,
+              itemList,
+              formModel,
+            ) ??
+            Column(children: itemList);
+      },
+    );
+  }
+}