+++ packages/reactive_forms_generator/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [0.11.0-beta]
+* widget for handling array of forms items see packages/reactive_forms_generator/example/lib/docs/mailing_list/delivery_route_form.dart for example of use
+
 ## [0.10.0-beta]
 * widget for handling array of items see packages/reactive_forms_generator/example/lib/docs/mailing_list/mailing_list_form.dart for example of use